requirements.txt
spacy
nltk
plotly
manim
numpy==1.26.4
protobuf>=3.20.3,<6.0.0dev
opencv-python
pillow
gtts
moviepy